pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grants-intelligence-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.133.0",
    "google-genai>=1.64.0",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "app"]
testpaths = ["app/tests"]
